chore: rearrange sinks to have DbSink initialized first
This will require some better API, given the HttpSink can
request un-synced events from database, it's better to have
the DbSink called first.

The order is important due to the implementation of the tracker,
which adds all the sinks into a list and calls them sequentially
when an event is received.