fix(android): switch to file based storage for events
Fixes #49

- Events are stored in events_log file. Each line of this file contains
an event serialized in the format expected by the server.
- Resource is serialized and stored in resource.json
- Some relational session data is still stored in sqlite like
session ID, start time, crashed and synced.
- To send data to server:
First event_log file is converted into events.json, which contains an
array of all events which were in event_log.
The transport layer sends the data to server using a single PUT request body.
There is support to use multipart request, but it will be enabled over time
as server supports it.