fix(webapp): use 'onboarded' flag in apps API response to set filter status
We were depending on the value of the filters API response to determine
if the selected app is onboarded or not in Overview page. This was a
brittle implementation leading to checks on each filter value to determine
"No data" state. This commit uses the 'onboarded' flag from apps API response
to accurately determine whether selected app is onboarded or not and update
the UI accordingly.