fix(webapp): fix invite member request failing
React state update weirdness was causing email and role fields
to not be updated when respective setState methods were called.
When invite api was called, this led to sending blank email
and default role in the request making it fail.

This state update failure seems to happen only in the case of form
fields. This commit fixes the issue by removing the form parent
and using a regular div instead as a parent to the email invite field,
role selector and invite button elements.