fix(webapp): add key id + timestamp only when pagination has actually occured
When first request to Crash/ANR group details returns a result, the
results array becomes non-empty. Any subsequent calls to the group details
API results in appending keyID and timestamp to the API regardless of whether
pagination has occured or not because there is no check for pagination actually
occuring and the only check is to see if the results array is non-empty.

This results in keyID and timestamp being sent when it shouldn't be. This
commit adds a check to see if pagination direction is not None which
guarantees that user has actually paginated before sending keyID and
timestamp.