chore(android): reduce calls to connectivity manager
NetworkStateAttributeProcessor needs the network state
each time an event is raised, so far this was calling
connectivity manager each time.

This change introduces a NetworkInfoProvider which
initializes with the network state on SDK
initialization and keeps up to date using the
listeners set in NetworkChangesCollector.

This also opens up cases where the most recent network
state needs to be queried.