fix(backend): use correct index when iterating exceptions and anrs
Fixes a bug which got uncovered when we started sending
exceptions and ANRs along with other events in a batch, earlier
exceptions and ANRs were being sent as a single event in a batch
which worked fine.

Also adds new sessions where a batch contains a mix of exception and other
events.