feat(android): improve session management
Earlier a session was created once and never updated
even if the app spent a long time in the background and
came back. This leads to sessions that can have gaps of
hours or days between event.

This commit ties the session duration to how a end user
might perceive it. That is, a new session is created when
an app is relaunched from background after a few seconds
(30 seconds by default).

This also has an effect on how app exit events are
collected as app exit is tracked by the OS and mapped to a
process ID. We need to map them back to sessions. Given
with the above change one pid can have multiple sessions,
we map the app exit event to the last session (by time)
for a particular pid.

This PR also changes when onAppForeground is called.
Earlier it was called after the lifecycle_* events were
tracked, however, it's now called before tracking the
lifecycle events. This ensures any initialization (like
for session manager) happens before the event is tracked.