fix(android): use process death time instead of current time for AppExit
The time at which the AppExit event is collected is not useful.
We're more interested in the time at which the app actually exited.

This also fixes a related issue in session replay where
session duration is calculated based on the time of the
last event. `AppExit` is typically the last event for a
session and using current time as it's timestamp makes
the session duration incorrect.

fixes: #953