chore(backend): remove eventIds array from crash + anr groups
closes #1027

Crash and ANR groups have a list of event ids associated with them
which are stored as an array in the 'event_ids' column in postgres.
Postgres fields have a limit of 1GB and this  list can get quite
large in high scale environments.

To improve this situation, this commit:
1. Removes the event_ids column from postgres
2. Fetches event ids of crash + anr groups from clickhouse based
on the event's fingeprint
3. Updates the ingest algorithm to no longer append event_ids to
crash & anr groups and only update updated_at and first_seen_event
timestamps
4. Updates the cleanup service to no longer delete event_ids
from postgres on stale event cleanup