fix(android): handle session management when elapsed time gets reset
We're using elapsed time since boot to store createdAt and lastEventTime
for recent session. This allows us to ensure that session management does
not get affected by any change in wall clock time.

However, we didn't handle the case where the device get rebooted and the
atomic clock is reset. This PR fixes that.

closes #1393