fix(frontend): place span durations above bar
fixes #1595

span duration display on bar had several issues:

- text overflowing the span bar
- text color change in selected state not being
visible in the overflowing part
- spans at end of trace not having enough space
to render text

In this commit, we render the text above the bar.
Now, text color does not need to change in selected
state and for spans near the end of the trace, text
becomes horizontally scrollable.