fix(frontend): filters not updating on no data/not onboarded states
fixes #2056

When filters api call returns a no data or not onboarded state,
we were not clearing previously set filters and also not updating
filters in the pages (due to changes being ignored because of
ready flag being false).

This commit clears filters on API failure status and applies
filter changes appropriately.