fix(frontend): select first root span available if url span invalid
fixes #2117

when user switches an app, new root spans are fetched and url is
updated. Before the new root spans are fetched, the url will
reflect the old root span selected. This would throw an error
since the old root span in url would not be valid for the new app.

This commit default to first available root span in case of invalid
url span instead of throwing an error.