fix(frontend): fix error on app name change
fixes #2145

when we change app name, we reload the page. Filters component
finds a persisted selected app which is pointing to the old name
and uses it instead of using the new app causing an error.

This commit checks uses the new app name correctly after app name
change to avoid the error.