fix(backend): fix exception grouping
When an event request contains multiple
exceptions with the same fingerprint the
function `bucketUnhandledExceptions` used
to return leading to any other fingerprints
to not get inserted into postgres table.

This issue shows up clearly with Flutter
because unhandled exceptions do not
necessarily crash & close the app.
So for example, if a button click is
leading to an unhandled exception, the
user may click it multiple times. Leading
to the multiple exceptions with same
fingerprint in the same event batch.

To fix this two changes have been made:
- remove the return statement when a new
exception group is created
- maintain an in-mem map of all fingerprints
processed in a batch to ensure no duplicate
exception groups are inserted in the db.