fix(frontend): check window before accessing persisted filters
Persisted filters are stored in session storage. When filters
component is evaluated by the next js server, it throws an error
because session storage is not available in node js. It's a browser
only feature.

This commit fixes the issue by checking if window exists before
attempting to fetch persisted filters.