chore(backend): add versioning column for bug reports table
closes #2312

Bug reports table in clickhouse did not have a versioning column.

This means that clikchouse will by default prioritise insert order.

When we query a bug report with an id and a FINAL clause it will
return what it thinks is the correct one which works fine when
we are not running distributed clickhouse but will break if we
are sharding/partitioning since insert order will depend on which
shard/partition we are talking to making it non-deterministic.

We fix this by explicity adding an `updated_at` column and using
that for versioning.

The steps are:
1. Create a new bug reports table (since we can't add versioning column
by altering a table)
2. Copy existing bug reports to it
3. Rename current bug reports table to `bug_reports_old`
4. Rename new table to `bug_reports`
5. Drop materialized view pointing to old table
6. Create materialized view pointing to new table
7. Drop `bug_reports_old`