fix(backend): show first available user ID in session timeline
fixes #2642

We were defaulting to User ID of the first event in the session.
In some sessions, the User ID may not be set at the time of the
first event resulting in an empty user ID even if the following
events have it.

This commit fixes this by keeping tracking of first available
User ID and setting it as the User ID for the session.