fix(backend): fix default user auth error on clickhouse 25.8 migration
fixes #2923

The issue was caused primarily by init script trying to grant all
permissions to admin user instead of the current grants it has
and was exarcebated by our custom users.xml file.

1. Clickhouse provides a way in docker compose to create a default user
through compose variables.If we use this, we don't need to have users.xml.
It creates a default user with ability to grant access to other users for us.
This commit removes users.xml so that it doesn't overwrite Clickhouse created
default user's config.
Doc link: https://clickhouse.com/docs/install#how-to-create-default-db-and-user
2. Default users can only grant the permissions it has. We were using "grant all"
in init script and sync scripts. This commit uses "grant current grants" so that
it can grant it's current permissions to admin user in both init and sync databases
 scripts.
3. Contributing guide was missing the step to run backfill script when migrating
from older codebase. This commit adds it in.